app/vinayak/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
pymysql
pyodbc
redis
pika
kafka-python
stomp.py
pymongo
python-dotenv
requests
pandas
httpx
python-multipart
alembic
yfinance
boto3
